+++ code/queuengin/src/main/java/com/neverwinterdp/queuengin/JSONMessageProducer.java
@@ -1,4 +1,4 @@
-package com.neverwinterdp.queue;
+package com.neverwinterdp.queuengin;
 
 import java.util.List;
