+++ code/queuengin/src/main/java/com/neverwinterdp/queuengin/StringMessageConsumerHandler.java
@@ -1,4 +1,4 @@
-package com.neverwinterdp.queue ;
+package com.neverwinterdp.queuengin ;
 
 
 public interface StringMessageConsumerHandler {
