+++ code/queuengin/src/main/java/com/neverwinterdp/queuengin/kafka/KafkaStringMessageProducer.java
@@ -1,4 +1,4 @@
-package com.neverwinterdp.queue.kafka;
+package com.neverwinterdp.queuengin.kafka;
 
 import java.util.ArrayList;
 import java.util.List;
