+++ code/queuengin/src/main/java/com/neverwinterdp/queuengin/kinesis/KinesisJSONMessageConsumerConnector.java
@@ -1,9 +1,9 @@
-package com.neverwinterdp.queue.kinesis ;
+package com.neverwinterdp.queuengin.kinesis ;
 
 import java.io.IOException;
 
-import com.neverwinterdp.queue.JSONMessageConsumerConnector;
-import com.neverwinterdp.queue.JSONMessageConsumerHandler;
+import com.neverwinterdp.queuengin.JSONMessageConsumerConnector;
+import com.neverwinterdp.queuengin.JSONMessageConsumerHandler;
 
 public class KinesisJSONMessageConsumerConnector<T> implements JSONMessageConsumerConnector<T>{
   public void consume(String topic, JSONMessageConsumerHandler<T> handler, int numOfThreads) throws IOException {
